Problem: Storybook does not work for Transaction component Solution: Transaction component refactored into TxView and Transaction container
