+++ src/components/tx/TxHistory/List/list.scss
@@ -1,5 +1,10 @@
 .columnName {
   text-transform: uppercase;
+  color: #8D8D8D !important;
+  font-size: 11px !important;
+  font-weight: 500 !important;
+  letter-spacing: 1px;
+  line-height: 16px;
 }
 
 .amountColumn {
