+++ stories/TxHistory/index.js
@@ -3,7 +3,7 @@ import { fromJS } from 'immutable';
 import { storiesOf } from '@storybook/react';
 // import { action } from '@storybook/addon-actions';
 import { MuiThemeProvider, getMuiTheme } from 'material-ui/styles';
-import { Transaction } from '../../src/components/tx/TxHistory/List/transaction';
+import { TxView } from '../../src/components/tx/TxHistory/List/TxView';
 
 const muiTheme = getMuiTheme({
   fontFamily: 'Rubik',
@@ -15,12 +15,12 @@ storiesOf('TxHistory', module)
       {story()}
     </MuiThemeProvider>
   ))
-  .add('Transaction', () => {
+  .add('TxView', () => {
     const tx = fromJS({
       from: '0x123',
       to: '0x123',
     });
     const fromAccount = fromJS({});
     const toAccount = fromJS({});
-    return (<Transaction tx={tx} fromAccount={fromAccount} toAccount={toAccount}/>);
+    return (<TxView tx={tx} fromAccount={fromAccount} toAccount={toAccount}/>);
   });
